convert fabric compute args to size_t instead of int

+++ mn/include/mn/Fabric.h
@@ -143,7 +143,7 @@ namespace mn
 
 	struct Compute_Dims
 	{
-		int x, y, z;
+		size_t x, y, z;
 	};
 
 	struct Compute_Args
